refactor: move these types to `types.d.ts`
